refactor: 로그인/회원가입 뷰모델 분리
- 회원가입/로그인 시 이전 상태를 저장하지 않기 위해 뷰 모델을 분리
