+++ lib/di/setup_providers.dart
@@ -66,5 +66,5 @@ List<SingleChildWidget> viewModels = [
   ChangeNotifierProvider<MapViewModel>(
       create: (context) => MapViewModel(context.read<UseCases>())),
   ChangeNotifierProvider<AuthViewModel>(
-      create: (context) => AuthViewModel(context.read<AuthRepository>()))
+      create: (context) => AuthViewModel(context.read<AuthRepository>())),
 ];
