+++ lib/presentation/auth/auth_event.dart
@@ -5,7 +5,5 @@ part 'auth_event.freezed.dart';
 
 @freezed
 abstract class AuthEvent with _$AuthEvent {
-  const factory AuthEvent.signUp(String email, String password, String repeatPassword, String nickname) = AuthSignUpEvent;
-  const factory AuthEvent.signIn(String email, String password) = AuthSignInEvent;
   const factory AuthEvent.signOut() = AuthSignOutEvent;
 }
