+++ lib/presentation/auth/local_sign_in/local_sign_in_sheet.dart
@@ -1,63 +1,63 @@
 import 'dart:async';
 
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
+import 'package:flutter_progress_hud/flutter_progress_hud.dart';
 import 'package:form_builder_validators/form_builder_validators.dart';
 import 'package:provider/provider.dart';
+import 'package:yeohaeng_ttukttak/data/repositories/auth_repository.dart';
 import 'package:yeohaeng_ttukttak/presentation/auth/auth_event.dart';
 import 'package:yeohaeng_ttukttak/presentation/auth/auth_view_model.dart';
 import 'package:yeohaeng_ttukttak/presentation/auth/components/form_errors.dart';
 import 'package:yeohaeng_ttukttak/presentation/auth/components/loading_dialog.dart';
-import 'package:yeohaeng_ttukttak/presentation/auth/components/local_sign_in_sheet.dart';
+import 'package:yeohaeng_ttukttak/presentation/auth/local_sign_in/local_sign_in_event.dart';
+import 'package:yeohaeng_ttukttak/presentation/auth/local_sign_in/local_sign_in_view_model.dart';
+import 'package:yeohaeng_ttukttak/presentation/auth/local_sign_up/local_sign_up_sheet.dart';
+import 'package:yeohaeng_ttukttak/presentation/auth/local_sign_up/local_sign_up_view_model.dart';
 import 'package:yeohaeng_ttukttak/presentation/map/map_screen.dart';
 import 'package:yeohaeng_ttukttak/presentation/welcome/welcome_screen.dart';
 
-class LocalSignUpSheet extends StatefulWidget {
-  const LocalSignUpSheet({super.key});
+class LocalSignInSheet extends StatefulWidget {
+  const LocalSignInSheet({super.key});
 
   @override
-  State<LocalSignUpSheet> createState() => _LocalSignUpSheetState();
+  State<LocalSignInSheet> createState() => _LocalSignInSheetState();
 }
 
-class _LocalSignUpSheetState extends State<LocalSignUpSheet> {
+class _LocalSignInSheetState extends State<LocalSignInSheet> {
   StreamSubscription? _subscription;
-
   final _formKey = GlobalKey<FormBuilderState>();
   final _emailFieldKey = GlobalKey<FormBuilderFieldState>();
   final _passwordFieldKey = GlobalKey<FormBuilderFieldState>();
-  final _repeatPasswordFieldKey = GlobalKey<FormBuilderFieldState>();
-  final _nicknameFieldKey = GlobalKey<FormBuilderFieldState>();
+
+  @override
+  void dispose() {
+    super.dispose();
+    FocusManager.instance.primaryFocus?.unfocus();
+    _subscription?.cancel();
+  }
 
   @override
   void initState() {
     super.initState();
 
     Future.microtask(() {
-      final viewModel = context.read<AuthViewModel>();
-      _subscription = viewModel.signUpStream.listen((event) => event.when(
+      final viewModel = context.read<LocalSignInViewModel>();
+      _subscription = viewModel.stream.listen((event) => event.when(
           showInputError: _onShowInputError,
           success: _onSuccess,
           loading: _onLoading));
     });
   }
 
-  void _onSuccess(String nickname) {
+  void _onSuccess() {
     while (Navigator.of(context).canPop()) {
       Navigator.of(context).pop();
     }
     Navigator.of(context).pushReplacement(
         MaterialPageRoute(builder: (context) => const MapScreen()));
-
-    _showSnackBar("$nickname님 가입을 축하합니다.");
-  }
-
-
-  void _showSnackBar(String message) {
-    final snackBar = SnackBar(
-        backgroundColor: Theme.of(context).colorScheme.secondaryContainer,
-        content: Text(message, style: Theme.of(context).textTheme.bodyLarge));
-
-    ScaffoldMessenger.of(context).showSnackBar(snackBar);
   }
 
   void _onShowInputError(target, message) {
@@ -67,23 +67,14 @@ class _LocalSignUpSheetState extends State<LocalSignUpSheet> {
   void _onLoading(bool isSubmitting) {
     showDialog(context: context, builder: (context) => const LoadingDialog());
 
-    Future.delayed(
-        const Duration(seconds: 1), () => {
-      if (mounted) Navigator.of(context).pop()
-    });
-  }
-
-  @override
-  void dispose() {
-    _subscription?.cancel();
-    FocusManager.instance.primaryFocus?.unfocus();
-    super.dispose();
+    Future.delayed(const Duration(seconds: 1),
+        () => {if (mounted) Navigator.of(context).pop()});
   }
 
   @override
   Widget build(BuildContext context) {
-    final viewModel = context.watch<AuthViewModel>();
-    final state = viewModel.signUpState;
+    final viewModel = context.watch<LocalSignInViewModel>();
+    final state = viewModel.state;
 
     return Padding(
       padding:
@@ -91,13 +82,12 @@ class _LocalSignUpSheetState extends State<LocalSignUpSheet> {
       child: SingleChildScrollView(
         child: FormBuilder(
           key: _formKey,
-          onChanged: () => _formKey.currentState?.save(),
           child: Container(
             padding: const EdgeInsets.fromLTRB(36, 36, 36, 60),
             width: MediaQuery.of(context).size.width,
             child:
                 Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
-              Text("이메일 회원가입",
+              Text("이메일 로그인",
                   style: Theme.of(context)
                       .textTheme
                       .titleLarge
@@ -107,10 +97,8 @@ class _LocalSignUpSheetState extends State<LocalSignUpSheet> {
                   key: _emailFieldKey,
                   name: 'email',
                   textInputAction: TextInputAction.next,
-                  validator: FormBuilderValidators.compose([
-                    FormBuilderValidators.required(),
-                    FormBuilderValidators.email()
-                  ]),
+                  validator: FormBuilderValidators.compose(
+                      [FormBuilderValidators.required()]),
                   decoration: const InputDecoration(
                       labelText: "이메일", hintText: "이메일(E-mail)을 입력하세요.")),
               const SizedBox(height: 28),
@@ -120,57 +108,12 @@ class _LocalSignUpSheetState extends State<LocalSignUpSheet> {
                   textInputAction: TextInputAction.next,
                   validator: FormBuilderValidators.compose([
                     FormBuilderValidators.required(),
-                    FormBuilderValidators.minLength(8),
-                    FormBuilderValidators.maxLength(64),
-                    (value) {
-                      if (value == value?.toLowerCase()) {
-                        return "대문자를 포함해야 합니다.";
-                      }
-                      return null;
-                    },
-                    FormBuilderValidators.match(".*\\d.*",
-                        errorText: "숫자를 포함해야 합니다."),
-                    FormBuilderValidators.match(
-                        r""".*[!@#$%^&*()_+\-=\[\]{};':"\\|,.<>\/?].*""",
-                        errorText: "특수 문자를 포함해야 합니다.")
                   ]),
                   obscureText: true,
                   decoration: const InputDecoration(
                     labelText: "비밀번호",
                     hintText: "비밀번호를 입력하세요.",
                   )),
-              const SizedBox(height: 28),
-              FormBuilderTextField(
-                  key: _repeatPasswordFieldKey,
-                  name: 'repeatPassword',
-                  obscureText: true,
-                  textInputAction: TextInputAction.next,
-                  decoration: const InputDecoration(
-                      labelText: "비밀번호 재입력", hintText: "비밀번호를 입력하세요."),
-                  validator: FormBuilderValidators.compose([
-                    FormBuilderValidators.equal(
-                        _formKey.currentState?.value['password'] ?? '',
-                        errorText: '입력된 두 비밀번호가 다릅니다.')
-                  ])),
-              const SizedBox(height: 28),
-              FormBuilderTextField(
-                  key: _nicknameFieldKey,
-                  name: 'nickname',
-                  textInputAction: TextInputAction.done,
-                  decoration: const InputDecoration(
-                      labelText: "닉네임", hintText: "닉네임을 입력하세요."),
-                  validator: FormBuilderValidators.compose([
-                    FormBuilderValidators.required(),
-                    FormBuilderValidators.minLength(3),
-                    FormBuilderValidators.maxLength(15),
-                    (value) {
-                      if (value != null &&
-                          !RegExp(r'/[가-힣a-zA-Z]/').hasMatch(value)) {
-                        return null;
-                      }
-                      return "반드시 한글 혹은 영문자를 포함해야 합니다.";
-                    }
-                  ])),
               const SizedBox(height: 24),
               FormErrors(errorMessages: state.errorMessages),
               OutlinedButton(
@@ -180,12 +123,8 @@ class _LocalSignUpSheetState extends State<LocalSignUpSheet> {
                     if (isValid == null || !isValid) return;
 
                     final values = _formKey.currentState?.value;
-
-                    viewModel.onEvent(AuthEvent.signUp(
-                        values?['email'],
-                        values?['password'],
-                        values?['repeatPassword'],
-                        values?['nickname']));
+                    viewModel.onEvent(LocalSignInEvent.signIn(
+                        values?['email'], values?['password']));
                   },
                   child: SizedBox(
                     width: MediaQuery.of(context).size.width,
@@ -194,7 +133,7 @@ class _LocalSignUpSheetState extends State<LocalSignUpSheet> {
                       mainAxisAlignment: MainAxisAlignment.spaceBetween,
                       children: [
                         Icon(Icons.email_outlined),
-                        Flexible(child: Center(child: Text("이메일로 회원가입"))),
+                        Flexible(child: Center(child: Text("이메일로 로그인"))),
                       ],
                     ),
                   )),
@@ -207,16 +146,19 @@ class _LocalSignUpSheetState extends State<LocalSignUpSheet> {
                         isScrollControlled: true,
                         useSafeArea: true,
                         context: context,
-                        builder: (context) => const LocalSignInSheet());
+                        builder: (context) => ChangeNotifierProvider(
+                            create: (context) => LocalSignUpViewModel(
+                                context.read<AuthRepository>()),
+                            child: const LocalSignUpSheet()));
                   },
                   child: RichText(
                       text: TextSpan(children: [
                     TextSpan(
-                        text: "이미 계정이 있으신가요? ",
+                        text: "처음 이신가요? ",
                         style: Theme.of(context).textTheme.bodyMedium?.copyWith(
                             color: Theme.of(context).colorScheme.outline)),
                     TextSpan(
-                        text: "로그인",
+                        text: "회원가입",
                         style: Theme.of(context).textTheme.bodyMedium)
                   ])),
                 ),
