+++ lib/presentation/auth/local_sign_in/local_sign_in_ui_event.freezed.dart
