+++ lib/presentation/auth/local_sign_up/local_sign_up_state.freezed.dart
@@ -3,7 +3,7 @@
 // ignore_for_file: type=lint
 // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
 
-part of 'local_sign_in_state.dart';
+part of 'local_sign_up_state.dart';
 
 // **************************************************************************
 // FreezedGenerator
@@ -14,33 +14,33 @@ T _$identity<T>(T value) => value;
 final _privateConstructorUsedError = UnsupportedError(
     'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
 
-LocalSignInState _$LocalSignInStateFromJson(Map<String, dynamic> json) {
-  return _LocalSignInState.fromJson(json);
+LocalSignUpState _$LocalSignUpStateFromJson(Map<String, dynamic> json) {
+  return _LocalSignUpState.fromJson(json);
 }
 
 /// @nodoc
-mixin _$LocalSignInState {
+mixin _$LocalSignUpState {
   List<String> get errorMessages => throw _privateConstructorUsedError;
 
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
   @JsonKey(ignore: true)
-  $LocalSignInStateCopyWith<LocalSignInState> get copyWith =>
+  $LocalSignUpStateCopyWith<LocalSignUpState> get copyWith =>
       throw _privateConstructorUsedError;
 }
 
 /// @nodoc
-abstract class $LocalSignInStateCopyWith<$Res> {
-  factory $LocalSignInStateCopyWith(
-          LocalSignInState value, $Res Function(LocalSignInState) then) =
-      _$LocalSignInStateCopyWithImpl<$Res, LocalSignInState>;
+abstract class $LocalSignUpStateCopyWith<$Res> {
+  factory $LocalSignUpStateCopyWith(
+          LocalSignUpState value, $Res Function(LocalSignUpState) then) =
+      _$LocalSignUpStateCopyWithImpl<$Res, LocalSignUpState>;
   @useResult
   $Res call({List<String> errorMessages});
 }
 
 /// @nodoc
-class _$LocalSignInStateCopyWithImpl<$Res, $Val extends LocalSignInState>
-    implements $LocalSignInStateCopyWith<$Res> {
-  _$LocalSignInStateCopyWithImpl(this._value, this._then);
+class _$LocalSignUpStateCopyWithImpl<$Res, $Val extends LocalSignUpState>
+    implements $LocalSignUpStateCopyWith<$Res> {
+  _$LocalSignUpStateCopyWithImpl(this._value, this._then);
 
   // ignore: unused_field
   final $Val _value;
@@ -62,22 +62,22 @@ class _$LocalSignInStateCopyWithImpl<$Res, $Val extends LocalSignInState>
 }
 
 /// @nodoc
-abstract class _$$LocalSignInStateImplCopyWith<$Res>
-    implements $LocalSignInStateCopyWith<$Res> {
-  factory _$$LocalSignInStateImplCopyWith(_$LocalSignInStateImpl value,
-          $Res Function(_$LocalSignInStateImpl) then) =
-      __$$LocalSignInStateImplCopyWithImpl<$Res>;
+abstract class _$$LocalSignUpStateImplCopyWith<$Res>
+    implements $LocalSignUpStateCopyWith<$Res> {
+  factory _$$LocalSignUpStateImplCopyWith(_$LocalSignUpStateImpl value,
+          $Res Function(_$LocalSignUpStateImpl) then) =
+      __$$LocalSignUpStateImplCopyWithImpl<$Res>;
   @override
   @useResult
   $Res call({List<String> errorMessages});
 }
 
 /// @nodoc
-class __$$LocalSignInStateImplCopyWithImpl<$Res>
-    extends _$LocalSignInStateCopyWithImpl<$Res, _$LocalSignInStateImpl>
-    implements _$$LocalSignInStateImplCopyWith<$Res> {
-  __$$LocalSignInStateImplCopyWithImpl(_$LocalSignInStateImpl _value,
-      $Res Function(_$LocalSignInStateImpl) _then)
+class __$$LocalSignUpStateImplCopyWithImpl<$Res>
+    extends _$LocalSignUpStateCopyWithImpl<$Res, _$LocalSignUpStateImpl>
+    implements _$$LocalSignUpStateImplCopyWith<$Res> {
+  __$$LocalSignUpStateImplCopyWithImpl(_$LocalSignUpStateImpl _value,
+      $Res Function(_$LocalSignUpStateImpl) _then)
       : super(_value, _then);
 
   @pragma('vm:prefer-inline')
@@ -85,7 +85,7 @@ class __$$LocalSignInStateImplCopyWithImpl<$Res>
   $Res call({
     Object? errorMessages = null,
   }) {
-    return _then(_$LocalSignInStateImpl(
+    return _then(_$LocalSignUpStateImpl(
       errorMessages: null == errorMessages
           ? _value._errorMessages
           : errorMessages // ignore: cast_nullable_to_non_nullable
@@ -96,15 +96,17 @@ class __$$LocalSignInStateImplCopyWithImpl<$Res>
 
 /// @nodoc
 @JsonSerializable()
-class _$LocalSignInStateImpl implements _LocalSignInState {
-  _$LocalSignInStateImpl({required final List<String> errorMessages})
+class _$LocalSignUpStateImpl implements _LocalSignUpState {
+  _$LocalSignUpStateImpl(
+      {final List<String> errorMessages = const [], this.verifyStarted = false})
       : _errorMessages = errorMessages;
 
-  factory _$LocalSignInStateImpl.fromJson(Map<String, dynamic> json) =>
-      _$$LocalSignInStateImplFromJson(json);
+  factory _$LocalSignUpStateImpl.fromJson(Map<String, dynamic> json) =>
+      _$$LocalSignUpStateImplFromJson(json);
 
   final List<String> _errorMessages;
   @override
+  @JsonKey()
   List<String> get errorMessages {
     if (_errorMessages is EqualUnmodifiableListView) return _errorMessages;
     // ignore: implicit_dynamic_type
@@ -113,14 +115,14 @@ class _$LocalSignInStateImpl implements _LocalSignInState {
 
   @override
   String toString() {
-    return 'LocalSignInState(errorMessages: $errorMessages)';
+    return 'LocalSignUpState(errorMessages: $errorMessages)';
   }
 
   @override
   bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
-            other is _$LocalSignInStateImpl &&
+            other is _$LocalSignUpStateImpl &&
             const DeepCollectionEquality()
                 .equals(other._errorMessages, _errorMessages));
   }
@@ -133,29 +135,30 @@ class _$LocalSignInStateImpl implements _LocalSignInState {
   @JsonKey(ignore: true)
   @override
   @pragma('vm:prefer-inline')
-  _$$LocalSignInStateImplCopyWith<_$LocalSignInStateImpl> get copyWith =>
-      __$$LocalSignInStateImplCopyWithImpl<_$LocalSignInStateImpl>(
+  _$$LocalSignUpStateImplCopyWith<_$LocalSignUpStateImpl> get copyWith =>
+      __$$LocalSignUpStateImplCopyWithImpl<_$LocalSignUpStateImpl>(
           this, _$identity);
 
   @override
   Map<String, dynamic> toJson() {
-    return _$$LocalSignInStateImplToJson(
+    return _$$LocalSignUpStateImplToJson(
       this,
     );
   }
 }
 
-abstract class _LocalSignInState implements LocalSignInState {
-  factory _LocalSignInState({required final List<String> errorMessages}) =
-      _$LocalSignInStateImpl;
+abstract class _LocalSignUpState implements LocalSignUpState {
+  factory _LocalSignUpState(
+      {final List<String> errorMessages,
+      final bool verifyStarted}) = _$LocalSignUpStateImpl;
 
-  factory _LocalSignInState.fromJson(Map<String, dynamic> json) =
-      _$LocalSignInStateImpl.fromJson;
+  factory _LocalSignUpState.fromJson(Map<String, dynamic> json) =
+      _$LocalSignUpStateImpl.fromJson;
 
   @override
   List<String> get errorMessages;
   @override
   @JsonKey(ignore: true)
-  _$$LocalSignInStateImplCopyWith<_$LocalSignInStateImpl> get copyWith =>
+  _$$LocalSignUpStateImplCopyWith<_$LocalSignUpStateImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }
