--- lib/presentation/auth/states/auth_state.freezed.dart
@@ -1,168 +1,0 @@
-// coverage:ignore-file
-// GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: type=lint
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
-
-part of 'auth_state.dart';
-
-// **************************************************************************
-// FreezedGenerator
-// **************************************************************************
-
-T _$identity<T>(T value) => value;
-
-final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
-
-AuthState _$AuthStateFromJson(Map<String, dynamic> json) {
-  return _AuthState.fromJson(json);
-}
-
-/// @nodoc
-mixin _$AuthState {
-  Member? get member => throw _privateConstructorUsedError;
-
-  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-  @JsonKey(ignore: true)
-  $AuthStateCopyWith<AuthState> get copyWith =>
-      throw _privateConstructorUsedError;
-}
-
-/// @nodoc
-abstract class $AuthStateCopyWith<$Res> {
-  factory $AuthStateCopyWith(AuthState value, $Res Function(AuthState) then) =
-      _$AuthStateCopyWithImpl<$Res, AuthState>;
-  @useResult
-  $Res call({Member? member});
-
-  $MemberCopyWith<$Res>? get member;
-}
-
-/// @nodoc
-class _$AuthStateCopyWithImpl<$Res, $Val extends AuthState>
-    implements $AuthStateCopyWith<$Res> {
-  _$AuthStateCopyWithImpl(this._value, this._then);
-
-  // ignore: unused_field
-  final $Val _value;
-  // ignore: unused_field
-  final $Res Function($Val) _then;
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? member = freezed,
-  }) {
-    return _then(_value.copyWith(
-      member: freezed == member
-          ? _value.member
-          : member // ignore: cast_nullable_to_non_nullable
-              as Member?,
-    ) as $Val);
-  }
-
-  @override
-  @pragma('vm:prefer-inline')
-  $MemberCopyWith<$Res>? get member {
-    if (_value.member == null) {
-      return null;
-    }
-
-    return $MemberCopyWith<$Res>(_value.member!, (value) {
-      return _then(_value.copyWith(member: value) as $Val);
-    });
-  }
-}
-
-/// @nodoc
-abstract class _$$AuthStateImplCopyWith<$Res>
-    implements $AuthStateCopyWith<$Res> {
-  factory _$$AuthStateImplCopyWith(
-          _$AuthStateImpl value, $Res Function(_$AuthStateImpl) then) =
-      __$$AuthStateImplCopyWithImpl<$Res>;
-  @override
-  @useResult
-  $Res call({Member? member});
-
-  @override
-  $MemberCopyWith<$Res>? get member;
-}
-
-/// @nodoc
-class __$$AuthStateImplCopyWithImpl<$Res>
-    extends _$AuthStateCopyWithImpl<$Res, _$AuthStateImpl>
-    implements _$$AuthStateImplCopyWith<$Res> {
-  __$$AuthStateImplCopyWithImpl(
-      _$AuthStateImpl _value, $Res Function(_$AuthStateImpl) _then)
-      : super(_value, _then);
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? member = freezed,
-  }) {
-    return _then(_$AuthStateImpl(
-      member: freezed == member
-          ? _value.member
-          : member // ignore: cast_nullable_to_non_nullable
-              as Member?,
-    ));
-  }
-}
-
-/// @nodoc
-@JsonSerializable()
-class _$AuthStateImpl implements _AuthState {
-  _$AuthStateImpl({this.member = null});
-
-  factory _$AuthStateImpl.fromJson(Map<String, dynamic> json) =>
-      _$$AuthStateImplFromJson(json);
-
-  @override
-  @JsonKey()
-  final Member? member;
-
-  @override
-  String toString() {
-    return 'AuthState(member: $member)';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$AuthStateImpl &&
-            (identical(other.member, member) || other.member == member));
-  }
-
-  @JsonKey(ignore: true)
-  @override
-  int get hashCode => Object.hash(runtimeType, member);
-
-  @JsonKey(ignore: true)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$AuthStateImplCopyWith<_$AuthStateImpl> get copyWith =>
-      __$$AuthStateImplCopyWithImpl<_$AuthStateImpl>(this, _$identity);
-
-  @override
-  Map<String, dynamic> toJson() {
-    return _$$AuthStateImplToJson(
-      this,
-    );
-  }
-}
-
-abstract class _AuthState implements AuthState {
-  factory _AuthState({final Member? member}) = _$AuthStateImpl;
-
-  factory _AuthState.fromJson(Map<String, dynamic> json) =
-      _$AuthStateImpl.fromJson;
-
-  @override
-  Member? get member;
-  @override
-  @JsonKey(ignore: true)
-  _$$AuthStateImplCopyWith<_$AuthStateImpl> get copyWith =>
-      throw _privateConstructorUsedError;
-}
